Okay. Good enough.
I definitely think that this is something I can hand in for the assignment.  Overall I changed the plotting functions (AGAIN) and made them to return RMSE. I was trying to make them plot the rmse in the graph, but this is intentionally left as an exercise for an interested reader.